Implemented and tested functions for inserting and removing from the list, both from the tail and the head.
Also added a function to test list emptiness.

Changed the prototypes of the functions to be more
pointer-centric as this makes using the functions and
macro much more consistent.

Instinct is to try and think of a means of allowing
users of the list to not have to use naked mallocs and frees,
but that is not the way of C...